requirements.txt
PyYAML==5.1
pytest==4.5.0
pytest-cov==2.7.1
python-interface==1.5.1
statistics==1.0.3.5
parameterized==0.7.0

#Openstack parser requirements
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr!=2.1.0,>=2.0.0 # Apache-2.0
Babel!=2.4.0,>=2.3.4 # BSD
cliff!=2.9.0,>=2.8.0 # Apache-2.0
python-dateutil>=2.5.3 # BSD
six>=1.10.0 # MIT
stevedore>=1.20.0 # Apache-2.0
requests>=2.14.2 # Apache-2.0